+++ src/main/java/com/googlecode/javaewah32/ChunkIteratorImpl32.java
@@ -19,11 +19,11 @@ import static com.googlecode.javaewah32.EWAHCompressedBitmap32.WORD_IN_BITS;
 final class ChunkIteratorImpl32 implements ChunkIterator {
 
     private final EWAHIterator32 ewahIter;
-    private final int sizeInBits;
+    private final long sizeInBits;
     private final Buffer buffer;
-    private int position;
+    private long position;
     private boolean runningBit;
-    private int runningLength;
+    private long runningLength;
     private int word;
     private int wordMask;
     private int wordPosition;
@@ -32,7 +32,7 @@ final class ChunkIteratorImpl32 implements ChunkIterator {
     private Boolean nextBit;
     private int nextLength;
 
-    ChunkIteratorImpl32(EWAHIterator32 ewahIter, int sizeInBits) {
+    ChunkIteratorImpl32(EWAHIterator32 ewahIter, long sizeInBits) {
         this.ewahIter = ewahIter;
         this.sizeInBits = sizeInBits;
         this.buffer = ewahIter.buffer();
@@ -134,7 +134,7 @@ final class ChunkIteratorImpl32 implements ChunkIterator {
     }
 
     private int runningOffset() {
-        return this.runningLength - this.position;
+        return (int) (this.runningLength - this.position);
     }
 
     private void movePosition(int offset) {
